Updated .Gitignore and Created the Hajime Discord Appplication
Database is no longer being committed to git. The data will be solely kept on the pi and only Yue will have access to the true data. Added a new discord application, Hajime, for development purposes. All updates will be made on Hajime before sending to git where Yue can use the updated code when needed. This was done because Yue is going to be running 24/7 now. Any changes to database columns will have to be manual for now. Hajime will use a separate db.

IMPORTANT
Need to make a file that updates the database columns for when changes are made. This can just be all of the database sync commands in one.